+++ R/AFM.rate.R
@@ -11,12 +11,10 @@
 #' @param IDs vector with AFM file IDs
 #' @param verbose if TRUE additional information is output
 #'
-#' @importFrom RAWdataR raw.getFileByID
 #' @importFrom utils write.csv read.csv
 #'
-#'
 #' @export
-AFM.rate <- function(dbFileName, IDs=NA, verbose = FALSE) {
+AFM.rate <- function(dbFileName, IDs=NA, fIDfile = NA, verbose = FALSE) {
   if (!file.exists(dbFileName)) stop("Cannot find DB file:", dbFileName)
   
   # load existing ratings
@@ -28,13 +26,12 @@ AFM.rate <- function(dbFileName, IDs=NA, verbose = FALSE) {
   IDs.all <- AFM.readDB(mydb)
   if (length(IDs)==0) {
     IDs <- IDs.all
-  } else {
-    IDs <- IDs[IDs %in% IDs.all]
-  }
+  } 
+  if (!is.na(fIDfile)) rID <- raw.readRAWIDfile(fIDfile) else rID = NULL
 
-  cat("\n\nRating AFM images.\n")
-  user.name <- readline("Last name of user: ")
-  
+  cat("\n\n--> Rating AFM images.\n")
+  user.name = 'anonymous'
+  if (interactive()) user.name <- readline("Last name of user: ")
   
   for(i in 1:length(IDs)) {
     # skip if this user has rated the image already
@@ -42,7 +39,17 @@ AFM.rate <- function(dbFileName, IDs=NA, verbose = FALSE) {
     if (length(m1)>0) next
     
     cat("Rating for ID=",IDs[i],"\n")
-    a = AFM.readDB(mydb, IDs[i])
+    if (IDs[i] %in% IDs.all) {
+      a = AFM.readDB(mydb, ID = IDs[i])
+    } else {
+      # try to read from local file
+      if (is.null(rID)) next
+      m <- which(rID$ID==IDs[i])
+      if (length(m)==0) next
+      fname = file.path(rID$path[m], rID$filename[m])
+      if (!file.exists(fname)) next
+      a = AFM.import(fname)
+    }
     a = AFM.flatten(a)
     print(plot(a))
     print(a)
@@ -54,5 +61,7 @@ AFM.rate <- function(dbFileName, IDs=NA, verbose = FALSE) {
   }
   DBI::dbDisconnect(mydb)
   
+  if (verbose) cat("Writing",nrow(df_ratings),"to SQLite databse.\n")
   AFM.writeRatings(dbFileName, df_ratings)
+  invisible(df_ratings)
 }