+++ R/AFM.writeDB.R
@@ -1,10 +1,9 @@
 #' Write AFM object to SQL DB
 #'
 #' @description
-#' AFMdata S4 image is written to an SQL database in the form
-#' of a data.frame; the ID will save it in a particular table
-#' in the database, if another AFM image with the same ID
-#' exists, it will be overwritten. 
+#' An AFMdata S4 image is written to an SQL database; the ID will 
+#' save it in a particular table in the database, if another AFM image 
+#' with the same ID exists, it will be overwritten. 
 #' 
 #' Open: mydb <- DBI::dbConnect(RSQLite::SQLite(), "myFile.sqlite")
 #'