+++ R/read.Park_header.R
@@ -78,6 +78,7 @@ loadBinaryDatafromTIFF <- function(fname) {
   close(to.read)
 
   # check for any reading errors
+  # this happens if the file size is odd
   if (length(which(is.na(q)==TRUE))>0) {
     warning(paste("reading error: NA found in",fname,"."))
     q[is.na(q)] <- 0